+++ example/android/app/src/main/kotlin/com/example/inappstory_plugin_example/MainActivity.kt
@@ -1,5 +1,5 @@
-package com.example.inappstory_plugin_example
+package com.inappstory.inappstory_plugin_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()