+++ ios/Classes/IASOnboardingsAdaptor.swift
@@ -1,33 +1,54 @@
-import Foundation
 import Flutter
-
+import Foundation
 @_spi(QAApp) import InAppStorySDK
 @_spi(IAS_API) import InAppStorySDK
 
-class IASOnboardingsAdaptor : IASOnboardingsHostApi {
-    init(binaryMessenger: FlutterBinaryMessenger, onboardingsAPI: OnboardingsAPI) {
+class IASOnboardingsAdaptor: IASOnboardingsHostApi {
+    init(
+        binaryMessenger: FlutterBinaryMessenger,
+        onboardingsAPI: OnboardingsAPI
+    ) {
         self.binaryMessenger = binaryMessenger
-        
+
         self.onboardingsAPI = onboardingsAPI
 
-        self.onboardingLoadCallbackFlutterApi = OnboardingLoadCallbackFlutterApi(binaryMessenger: binaryMessenger);
-        
-        IASOnboardingsHostApiSetup.setUp(binaryMessenger: binaryMessenger, api: self)
+        self.onboardingLoadCallbackFlutterApi =
+            OnboardingLoadCallbackFlutterApi(binaryMessenger: binaryMessenger)
+
+        IASOnboardingsHostApiSetup.setUp(
+            binaryMessenger: binaryMessenger,
+            api: self
+        )
     }
-    
+
     private var binaryMessenger: FlutterBinaryMessenger
-    
+
     private var onboardingsAPI: OnboardingsAPI
-    
-    private var onboardingLoadCallbackFlutterApi: OnboardingLoadCallbackFlutterApi
-    
+
+    private var onboardingLoadCallbackFlutterApi:
+        OnboardingLoadCallbackFlutterApi
+
     func show(limit: Int64, feed: String, tags: [String]) throws {
         func complete(show: Bool) {
-            if (show) {
-                onboardingLoadCallbackFlutterApi.onboardingLoad(count: limit, feed: feed) { _ in }
+            if show {
+                onboardingLoadCallbackFlutterApi.onboardingLoadSuccess(
+                    count: limit,
+                    feed: feed
+                ) { _ in }
+            } else {
+                onboardingLoadCallbackFlutterApi.onboardingLoadError(
+                    feed: feed,
+                    reason: "Onboarding load error"
+                ) { _ in }
             }
         }
-        
-        onboardingsAPI.showOnboarding(feed: feed, limit: Int(limit), with:tags, with:InAppStory.shared.panelSettings ,complete: complete)
+
+        onboardingsAPI.showOnboarding(
+            feed: feed,
+            limit: Int(limit),
+            with: tags,
+            with: InAppStory.shared.panelSettings,
+            complete: complete
+        )
     }
 }